fix(tabs-extended-media): Design QA fixes.
