+++ packages/web-components/src/components/tabs-extended-media/__stories__/README.stories.mdx
@@ -1,0 +1,70 @@
+import { Props, Description } from '@storybook/addon-docs/blocks';
+import contributing from '../../../../../../docs/contributing-license.md';
+import markdownContents from '../../../../docs/markdown-contents.md';
+
+# `<dds-tabs-extended-media>`
+
+> The Tabs extended with media - Tabbed collection of `<dds-tab>` content items.
+
+> 💡 Check our
+> [CodeSandbox](https://githubbox.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/web-components/examples/codesandbox/components/tabs-extended-media)
+> example implementation.
+
+[![Edit @carbon/ibmdotcom-web-components](https://codesandbox.io/static/img/play-codesandbox.svg)](https://githubbox.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/web-components/examples/codesandbox/components/tabs-extended-media)
+
+## Getting started
+
+Here's a quick example to get you started.
+
+##### JS
+
+```javascript
+import '@carbon/ibmdotcom-web-components/es/components/tabs-extended-media/tabs-extended-media.js';
+```
+
+##### HTML
+
+```html
+<dds-tabs-extended-media>
+  <dds-content-section-heading>Section heading</dds-content-section-heading>
+  <dds-tab label="First tab">
+    <dds-content-item-horizontal-media align="right">
+      <dds-image
+        slot="media"
+        alt="Lorem ipsum"
+        default-src="https://fpoimg.com/672x336?text=2:1&bg_color=ee5396&text_color=161616"
+      ></dds-image>
+      <dds-content-item-heading>Aliquam condimentum</dds-content-item-heading>
+      <dds-content-item-horizontal-copy
+        >Lorem ipsum dolor sit amet, _consectetur_ adipiscing elit. Aenean et ultricies est. Mauris iaculis eget dolor nec
+        hendrerit. Phasellus at elit sollicitudin.</dds-content-item-horizontal-copy
+      >
+      <dds-link-list slot="footer" type="vertical">
+        <dds-link-list-item-cta icon-placement="right" href="https://www.ibm.com" cta-type="local">
+          Learn more
+        </dds-link-list-item-cta>
+        <dds-link-list-item-cta icon-placement="right" href="https://www.ibm.com" cta-type="external">
+          Microservices and containers
+        </dds-link-list-item-cta>
+      </dds-link-list>
+    </dds-content-item-horizontal-media>
+  </dds-tab>
+</dds-tabs-extended-media>
+```
+
+## Props
+
+### `<dds-tab>`
+
+<Props of="dds-tab" />
+<Description markdown={markdownContents} />
+
+## Stable selectors
+
+See [our README](https://github.com/carbon-design-system/carbon-for-ibm-dotcom/tree/master/packages/web-components#stable-selectors-for-analytics-and-integratione2e-testing-in-web-components) to see how Web Components selector and `data-autoid` should be used.
+
+| Web Components selector     | Compatibility selector                   | Description |
+| --------------------------- | ---------------------------------------- | ----------- |
+| `<dds-tabs-extended-media>` | `data-autoid="dds--tabs-extended-media"` | Component   |
+
+<Description markdown={contributing} />
