+++ packages/web-components/src/components/tabs-extended-media/tabs-extended-media.scss
